Fix: phpstan error

+++ src/helpers.php
@@ -299,14 +299,14 @@ function seconds_to_hours_str( $seconds ) {
 	if ( ! empty( $hours ) ) {
 		$duration_str .= sprintf(
 			// phpcs:ignore WordPress.WP.I18n.MissingTranslatorsComment
-			_n( '%s Hour', '%s Hours', $hours, 'bluedolphin-lms' ),
+			_n( '%s Hour', '%s Hours', (int) $hours, 'bluedolphin-lms' ),
 			number_format_i18n( $hours )
 		);
 	}
 
 	$mins = $seconds % 3600;
 	if ( ! empty( $mins ) ) {
-		$mins          = gmdate( 'i', $mins );
+		$mins          = (int) gmdate( 'i', $mins );
 		$duration_str .= sprintf(
 			// phpcs:ignore WordPress.WP.I18n.MissingTranslatorsComment
 			_n( ' %s Min', ' %s Mins', $mins, 'bluedolphin-lms' ),
